% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censo_poblacion.R
\name{censo_poblacion}
\alias{censo_poblacion}
\title{Censos y Conteos de Población y Vivienda.}
\usage{
censo_poblacion(type = "iter", year = "2010", state = "Nacional",
  totalent = FALSE, totalmun = FALSE, noloc = FALSE,
  totalageb = FALSE, muestra = NA)
}
\arguments{
\item{type}{Proyecto INEGI. Opciones: "iter", "ageb", "EntornoUrbano", "Rural", y "Muestra". La opción "Muestra" tiene argumentos adicionales.}

\item{year}{Año requerido del censo o conteo. En esta versión están disponibles 2000, 2005, 2010 y 2015.}

\item{state}{Descarga datos por estado. Utilizar el nombre del estado, sin espacios. Ejemplos: "Aguascalientes", "CDMX", "SanLuisPotosi". NA/Default = Nacional.}

\item{totalent}{Para el formato "iter", TRUE para obtener, además de la información a nivel localidad, resumen estadístico a nivel estado.}

\item{totalmun}{Para el formato "iter", TRUE para obtener, además de la información a nivel localidad, resumen estadístico a nivel municipio.}

\item{noloc}{Para el formato "iter", TRUE para no obtener información observaciones a nivel localidad.}

\item{totalageb}{Para el formato "ageb", TRUE para obtener, además de la información a nivel manzana, resumen estadístico a nivel ageb.}

\item{muestra}{Para el formato "Muestra", utilizar "Migrantes", "Personas", o "Viviendas" según el tipo de base de datos requerida.}
}
\description{
Censos y Conteos de Población y Vivienda.
}
