% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censo_poblacion_ageb.R
\name{censo_poblacion_ageb}
\alias{censo_poblacion_ageb}
\title{Censos y Conteos de Población y Vivienda.
AGEB}
\usage{
censo_poblacion_ageb(year = "2010", estado = NA, totalestado = FALSE,
  totalmunici = FALSE, nolocalidad = FALSE, totalageb = FALSE)
}
\arguments{
\item{year}{Año requerido del censo o conteo. En esta versión están disponibles 2000, 2005, 2010 y 2015.}

\item{estado}{Descarga datos por estado. Utilizar el nombre del estado, con espacios. Ejemplos: "Aguascalientes", "CDMX", "San Luis Potosi".}

\item{totalestado}{TRUE para obtener, además de la información a nivel localidad, resumen estadístico a nivel estado.}

\item{totalmunici}{TRUE para obtener, además de la información a nivel localidad, resumen estadístico a nivel municipio.}

\item{nolocalidad}{TRUE para no obtener información observaciones a nivel localidad.}

\item{totalageb}{TRUE para obtener, además de la información a nivel manzana, resumen estadístico a nivel ageb.}
}
\description{
Censos y Conteos de Población y Vivienda.
AGEB
}
