% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censo_poblacion_muestra.R
\name{censo_poblacion_muestra}
\alias{censo_poblacion_muestra}
\title{Censos y Conteos de Población y Vivienda.
MUESTRA}
\usage{
censo_poblacion_muestra(year = "2010", estado = NA, muestra = NA)
}
\arguments{
\item{year}{Año requerido del censo o conteo. En esta versión están disponibles 2000, 2005, 2010 y 2015.}

\item{estado}{Descarga datos por estado. Utilizar el nombre del estado, con espacios. Ejemplos: "Aguascalientes", "CDMX", "San Luis Potosi".}

\item{muestra}{Utilizar "Migrantes", "Personas", o "Viviendas" según el tipo de base de datos requerida.}
}
\description{
Censos y Conteos de Población y Vivienda.
MUESTRA
}
