% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censo_poblacion_rural.R
\name{censo_poblacion_rural}
\alias{censo_poblacion_rural}
\title{Censos y Conteos de Población y Vivienda.
Resultados sobre localidades con menos de 5 mil habitantes}
\usage{
censo_poblacion_rural(year = "2010", estado = "Nacional",
  totalestado = FALSE, totalmunici = FALSE, nolocalidad = FALSE)
}
\arguments{
\item{year}{Año requerido del censo o conteo. En esta versión están disponibles 2000, 2005, 2010 y 2015.}

\item{estado}{Descarga datos por estado. Utilizar el nombre del estado, con espacios. Ejemplos: "Aguascalientes", "CDMX", "San Luis Potosi", "Nacional".}

\item{totalestado}{TRUE para obtener, además de la información a nivel localidad, resumen estadístico a nivel estado.}

\item{totalmunici}{TRUE para obtener, además de la información a nivel localidad, resumen estadístico a nivel municipio.}

\item{nolocalidad}{TRUE para no obtener información observaciones a nivel localidad.}
}
\description{
Censos y Conteos de Población y Vivienda.
Resultados sobre localidades con menos de 5 mil habitantes
}
