% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enigh.R
\name{enigh}
\alias{enigh}
\title{ENIGH Nueva Construccion (2018-2022)}
\usage{
enigh(year = NA, datos = "", formato = "dbf", extdir = "")
}
\arguments{
\item{year}{Año de levantamiento de la encuesta en formato numerico. Los años disponibles son 2018, 2020, 2022}

\item{datos}{Base de datos a descargar "viviendas" "hogares" "concentrado" "erogaciones" "gastohogar" "gastotarjetas" "poblacion" "ingresos" "gastopersona" "trabajos" "agro" "noagro"}

\item{formato}{Formato del archivo, 'csv' 'dbf' 'sav' 'dta'}

\item{extdir}{Si se descomprimirá en algún directorio externo}
}
\value{
Data.frame
}
\description{
Descarga datos de la Encuesta Nacional de Ingreso y Gasto de los Hogares, Nueva Construccion (2018-2022).
}
\details{
La ENIGH provee informacion estadisticas sobre los ingresos y gastos de los hogares en cuanto a su monto, procedencia y distribucion. Adicionalmente, la ENIGH provee informacion sobre las caracteristicas socio-demograficas de los integrantes del hogar.
}
\examples{

# Descargar datos de hogares
\dontrun{pob = enigh(year = 2022,  datos = "poblacion")}
}
