% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enigh_nueva.R
\name{enigh_nueva}
\alias{enigh_nueva}
\title{ENIGH Nueva Construccion (2008-2014)}
\usage{
enigh_nueva(year = NA, hogares = FALSE, concentrado = FALSE,
  erogaciones = FALSE, gastohogar = FALSE, gastotarjetas = FALSE,
  poblacion = FALSE, ingresos = FALSE, gastopersona = FALSE,
  trabajos = FALSE, agro = FALSE, noagro = FALSE)
}
\arguments{
\item{year}{Año de levantamiento de la encuesta}

\item{hogares}{En esta tabla se encuentran contenidas las características de los hogares que habitan los integrantes de los mismos.}

\item{concentrado}{Principales variables por hogar. En esta tabla se encuentran las variables construidas a partir de las otras tablas de la base de datos. Registra el resumen concentrado por hogar, de ingresos y gastos en toda modalidad posible. Además todos los ingresos y gastos registrados en esta tabla son trimestrales.}

\item{erogaciones}{Esta tabla permite identificar las erogaciones financieras y de capital realizadas por hogar.}

\item{gastohogar}{Esta tabla permite identificar los gastos monetarios y no monetarios del hogar.}

\item{gastotarjetas}{Esta tabla permite identificar en qué medida los hogares se auxilian del financiamiento que les ofrece la tarjeta de crédito bancaria o comercial para cubrir algunas necesidades básicas.}

\item{poblacion}{Esta tabla proviene de la tabla HOGARES. Identifica las características sociodemográficas de los integrantes del hogar; y el acceso a las instituciones de salud que tienen los integrantes del hogar.}

\item{ingresos}{Esta tabla permite identificar los ingresos y percepciones financieras y de capital de cada uno de los integrantes del hogar, por diversos conceptos.}

\item{gastopersona}{Esta tabla permite identificar los gastos realizados por cada integrante del hogar en educación, transporte público y remuneraciones en especie.}

\item{trabajos}{Esta tabla muestra la condición de actividad de los integrantes del hogar de 12 o más años y algunas características ocupacionales durante el periodo de referencia.}

\item{agro}{Esta tabla muestra la información de los trabajadores independientes, mayores de 12 años que tienen en el hogar negocios dedicados a las actividades agrícolas, forestales y de tala, además de actividades de cría, explotación y productos derivados de la pesca y caza.}

\item{noagro}{Esta tabla muestra a los trabajadores independientes mayores de 12 años, que tienen negocios en el hogar dedicados a las actividades industriales, comerciales y de servicios.}
}
\description{
ENIGH Nueva Construccion (2008-2014)
}
