% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_caminos.R
\name{sig_caminos}
\alias{sig_caminos}
\title{Red Nacional de Caminos}
\usage{
sig_caminos(year = NA)
}
\arguments{
\item{year}{An`o de la version de la Red Nacional de Caminos (disponibles 2016, 2017, 2018)}
}
\description{
Red Nacional de Caminos
}
