% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_caminos_descarga.R
\name{sig_caminos_descarga}
\alias{sig_caminos_descarga}
\title{Red Nacional de Caminos - Descarga Datos}
\usage{
sig_caminos_descarga(year = NA)
}
\arguments{
\item{year}{Año de referencia del mapa, en formato numerico (2014-2018).}
}
\description{
#' Esta funcion descarga un una lista todos los Shapefiles de la Red Nacional de Caminos para un año especifico. El objeto resultante de esta funcion es necesario para extraer cada Shapefile con la funcion \code{sig_caminos_extrae()}.
}
