% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_caminos_extrae.R
\name{sig_caminos_extrae}
\alias{sig_caminos_extrae}
\title{Red Nacional de Caminos II}
\usage{
sig_caminos_extrae(. = NA, year = NA, conjunto = NA)
}
\arguments{
\item{.}{Objeto (inserta el objeto creado con sig_caminos())}

\item{year}{An~o de la version de la Red Nacional de Caminos (disponibles 2014-2018)}

\item{conjunto}{Capa de Shapefile a descargar. Disponibles: estructura, localidad, maniobra_prohibida, plaza_cobro, poste_de_referencia, puente, red_vial, sitio_de_interes, tarifas, transbordador, tred_localidad, tred_sitio_de_interes, runion.}
}
\description{
Red Nacional de Caminos II
}
