% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogo_inegi.R
\name{catalogo_inegi}
\alias{catalogo_inegi}
\title{Cataologo INEGI}
\usage{
catalogo_inegi(id = NA)
}
\arguments{
\item{id}{Numero de identificacion unico de las bases de datos disponibles en el repositorio la Red Nacional de Metadatos del INEGI}
}
\description{
Consulta el catalogo la Red Nacional de Metadatos del INEGI o los metadatos de una base de datos en particular.
}
\details{
La funcion \code{catalogo_inegi} provee una lista de bases de datos con un identificador unico (\code{id}). Si conoces el \code{id} de la base de datos a consultar, utilizalo en el parametro para obtener el libro de codigos y los metadatos de la base de datos. Si no conoces el \code{id} de la base de datos a consultar, teclea la funcion \code{catalogo_inegi} sin parametros para descargar la lista de bases de datos (Ver ejemplo abajo).
}
\examples{

# Almacena el catalogo como base de datos.
dt.catalogo <- catalogo_inegi()

# Consulta metadatos de una base de datos.
catalogo_inegi(id = 489)

}
\seealso{
Consulta el \href{http://www3.inegi.org.mx/rnm/index.php/catalog/}{repositorio la Red Nacional de Metadatos del INEGI.}
}
