% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censo_municipal.R
\name{censo_municipal}
\alias{censo_municipal}
\title{Microdatos del Censo Nacional de Gobiernos Municipales y Delegacionales}
\usage{
censo_municipal(fuente = NA, year = NA, datos = NA)
}
\arguments{
\item{fuente}{Fuente de datos de las instituciones públicas de municipales y delegacionales en formato alfanumerico. Las opciones son: ayuntamiento, administracion, seguridad, justicia.}

\item{year}{Año del levantamiento del censo en formato numerico. Los años disponibles son 2011, 2013, 2015 y 2017.}

\item{datos}{Base de datos producida por cada fuente de datos en formato alfanumerico. Las opciones pueden ser, segun la fuente de datos: comision, estructura, funciones, marco, participacion, recursos, tramites, transparencia, ejercicio, infraestructura, y recursos.}
}
\description{
Descarga los microdatos del Censo Nacional de Gobiernos Municipales y Delegacionales (2011, 2013, 2015 y 2017).
}
\details{
Los microdatos del censo contienen información de 2,454 gobiernos municipales y gobiernos delegacionales de la CDMX. Si conoces el \code{id} de la base de datos a consultar, utilizalo en el parametro para obtener el libro de codigos y los metadatos de la base de datos. Si no conoces el \code{id} de la base de datos a consultar, teclea la funcion \code{catalogo_inegi} sin parametros para descargar la lista de bases de datos (Ver ejemplo abajo).
}
\examples{

# Consulta los metadatos del Censo Nacional de Gobiernos Municipales y Delegacionales
\dontrun{censo_municipal()}

# Descarga los microdatos de la estructura de los ayuntamientos en 2011
\dontrun{dt.estructura2011 <- censo_municipal(fuente = "ayuntamiento", "estructura", year = 2011)}

}
\seealso{
Consulta la relacion entre fuentes y datos en el articulo \href{link}{TODO::VIGNETTE}.
}
