% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censo_poblacion_ageb.R
\name{censo_poblacion_ageb}
\alias{censo_poblacion_ageb}
\title{Censo de Poblacion - AGEB}
\usage{
censo_poblacion_ageb(year = "2010", estado = NA, totalestado = FALSE,
  totalmunicipio = FALSE, totalageb = FALSE)
}
\arguments{
\item{year}{Año del levantamiento del censo en formato numerico. Los años disponibles (incluyendo los conteos) son: 2000, 2005, 2010 y 2015.}

\item{estado}{Define el nombre de la entidad federativa para descargar los datos, en formato alfanumerico. Utiliza "Nacional" para descargarlos a nivel nacional. Los nombres de los estados deben ir capitalizados (y en su caso, con espacios), por ejemplo: "Aguascalientes", "CDMX", "San Luis Potosi".}

\item{totalestado}{Resultados agregados a nivel entidad federativa. \code{FALSE} omite los resultados a nivel entidad federativa.}

\item{totalmunicipio}{Resultados agregados a nivel municipio. \code{FALSE} omite los resultados a nivel municipio.}

\item{totalageb}{Resultados agregados a nivel AGEB. \code{FALSE} omite los resultados a nivel AGEB.}
}
\description{
Descarga los datos del Censo de Poblacion y Vivienda al nivel de desagregacion AGEB y manzana urbana.
}
\details{
Esta base de datos tiene tres niveles de agregacion: entidades federativas, municipios, agebs y manzanas (en zonas urbanas).
}
\examples{

# Consulta los datos del Censo de Poblacion y Vivienda a nivel AGEB y manzana urbana.
censo_poblacion_ageb()

# Descarga los datos de San Luis Potosi de 2010.
dt.ageb.sanluis2010 = censo_poblacion_ageb(year = 2010, estado = "San Luis Potosi")

}
\concept{conteo_poblacion_ageb()}
