% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censo_poblacion_muestra.R
\name{censo_poblacion_muestra}
\alias{censo_poblacion_muestra}
\title{Censo de Poblacion - Muestra}
\usage{
censo_poblacion_muestra(year = "2010", estado = NA, muestra = NA)
}
\arguments{
\item{year}{Año del levantamiento del censo en formato numerico. Los años disponibles (incluyendo los conteos) son: XXXXXXX.}

\item{estado}{Define el nombre de la entidad federativa para descargar los datos, en formato alfanumerico. La funcion, por defecto utiliza la palabra "Nacional" para descargar los datos de todos los estados. Los nombres de los estados deben ir capitalizados (y en su caso, con espacios), por ejemplo: "Aguascalientes", "CDMX", "San Luis Potosi".}

\item{muestra}{Utilizar "Migrantes", "Personas", o "Viviendas" según el tipo de base de datos requerida.}
}
\description{
Censo de Poblacion y Vivienda. Muestra (cuestionario ampliado)
}
\details{
Esta base de datos tiene dos niveles de agregacion: entidades federativas y municipios.
}
\examples{

# Consulta los metadatos de la muestra del Censo de Poblacion y Vivienda
censo_poblacion_muestra()

# Descarga los datos de San Luis Potosi de 2010.
dt.muestra.sanluis2010 = censo_poblacion_muestra(year = 2010, estado = "San Luis Potosi", muestra = "Personas")

}
\concept{conteo_poblacion_muestra()}
