% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censo_poblacion_rural.R
\name{censo_poblacion_rural}
\alias{censo_poblacion_rural}
\title{Censo de Poblacion - Localidades Rurales}
\usage{
censo_poblacion_rural(year = "2010", estado = "Nacional",
  totalestado = FALSE)
}
\arguments{
\item{year}{Año del levantamiento del censo en formato numerico. Los años disponibles (incluyendo los conteos) son: 1990, 2000, 2005, 2010 y 2015.}

\item{estado}{Define el nombre de la entidad federativa para descargar los datos, en formato alfanumerico. La funcion, por defecto utiliza la palabra "Nacional" para descargar los datos de todos los estados. Los nombres de los estados deben ir capitalizados (y en su caso, con espacios), por ejemplo: "Aguascalientes", "CDMX", "San Luis Potosi".}

\item{totalestado}{Resultados agregados a nivel entidad federativa. \code{FALSE} omite los resultados a nivel entidad federativa}
}
\description{
Censo de Poblacion y Vivienda. Resultados sobre localidades con menos de 5 mil habitantes
}
\details{
Esta base de datos tiene dos niveles de agregacion: entidades federativas y municipios.
}
\examples{

# Consulta los datos sobre localidades rurales del Censo de Poblacion y Vivienda
censo_poblacion_rural()

# Descarga los datos de San Luis Potosi de 2010.
dt.rural.sanluis2010 = censo_poblacion_rural(year = 2010, estado = "San Luis Potosi")

}
\concept{conteo_poblacion_rural()}
