% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censo_poblacion_entorno.R
\name{censo_poblacion_urbano}
\alias{censo_poblacion_urbano}
\title{Censos y Conteos de Población y Vivienda.
ENTORNO}
\usage{
censo_poblacion_urbano(year = "2010", estado = NA,
  totalestado = FALSE, totalmunici = FALSE, nolocalidad = FALSE)
}
\arguments{
\item{year}{Año requerido del censo o conteo. En esta versión están disponibles 2000, 2005, 2010 y 2015.}

\item{estado}{Descarga datos por estado. Utilizar el nombre del estado, con espacios. Ejemplos: "Aguascalientes", "CDMX", "San Luis Potosi".}

\item{totalestado}{TRUE para obtener, además de la información a nivel localidad, resumen estadístico a nivel estado.}

\item{totalmunici}{TRUE para obtener, además de la información a nivel localidad, resumen estadístico a nivel municipio.}

\item{nolocalidad}{TRUE para no obtener información observaciones a nivel localidad.}
}
\description{
Censos y Conteos de Población y Vivienda.
ENTORNO
}
