% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enigh_nuevaconstruccion.R
\name{enigh_nuevaconstruccion}
\alias{enigh_nuevaconstruccion}
\title{ENIGH Nueva Construccion (2008-2014)}
\usage{
enigh_nuevaconstruccion(year = NA, hogares = FALSE,
  concentrado = FALSE, erogaciones = FALSE, gastohogar = FALSE,
  gastotarjetas = FALSE, poblacion = FALSE, ingresos = FALSE,
  gastopersona = FALSE, trabajos = FALSE, agro = FALSE,
  noagro = FALSE)
}
\arguments{
\item{year}{Año de levantamiento de la encuesta en formato numerico. Los años disponibles son 2008, 2010, 2012 y 2014}

\item{hogares}{Características de los hogares que habitan los integrantes de los mismos.}

\item{concentrado}{Principales variables por hogar. Contiene un resumen de variables de otras tablas.}

\item{erogaciones}{Erogaciones financieras y de capital realizadas por hogar.}

\item{gastohogar}{Gastos monetarios y no monetarios del hogar.}

\item{gastotarjetas}{Financiamiento de tarjeta de crédito bancaria o comercial.}

\item{poblacion}{Características sociodemográficas de los integrantes del hogar.}

\item{ingresos}{Ingresos y percepciones financieras y de capital de cada uno de los integrantes del hogar, por diversos conceptos.}

\item{gastopersona}{Gastos realizados por cada integrante del hogar en educación, transporte público y remuneraciones en especie.}

\item{trabajos}{Actividad de los integrantes del hogar de 12 o más años.}

\item{agro}{Información de los trabajadores independientes dedicados a las actividades agrícolas, pesca, caza, etc.}

\item{noagro}{Informacion de los trabajadores independientes que tienen negocios en el hogar dedicados a las actividades industriales, comerciales y de servicios.}
}
\description{
Encuesta Nacional de Ingreso y Gasto de los Hogares (ENIGH). Nueva Construccion (2008-2014).
}
