% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enoe.R
\name{enoe}
\alias{enoe}
\title{Title}
\usage{
enoe(year = NA, trimestre = NA, integrar = FALSE)
}
\arguments{
\item{year}{Año de levantamiento de la encuesta}

\item{trimestre}{Opciones.}

\item{integrar}{Text.}
}
\description{
Title
}
