% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_caminos_descarga.R
\name{sig_caminos_descarga}
\alias{sig_caminos_descarga}
\title{Red Nacional de Caminos - Descarga Datos}
\usage{
sig_caminos_descarga(year = NA)
}
\arguments{
\item{year}{An`o de la version de la Red Nacional de Caminos (disponibles 2016, 2017, 2018)}
}
\description{
Red Nacional de Caminos - Descarga Datos
}
