% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_marcogeo.R
\name{sig_marcogeo}
\alias{sig_marcogeo}
\title{Marco Geoestadistico Nacional}
\usage{
sig_marcogeo(year = NA, capa = NA, version = NA)
}
\arguments{
\item{year}{An~o (1995-2018)}

\item{capa}{Opciones: entidades, municipios, ageb, urbano, rural}

\item{version}{Algunos an~ tienen multiples versiones. Si no la sabes, el mensaje de error te dara las opcines.}
}
\description{
Marco Geoestadistico Nacional
}
